RoCEv2 tests: retire relocated-codec tests, gate C++ tests on NO_ROCEV2
- Delete test_rocev2_corner_cases.py and slim test_rocev2.py / conftest.py:
  the metadata-bus codec they exercised now lives in surf, so the codec/QP
  state-machine tests are no longer this repo's responsibility.
- Gate the C++ rocev2 test subdir on (NOT NO_ROCEV2) so a -DNO_ROCEV2 build
  does not try to compile tests against absent ibverbs sources.
