+++ tests/protocols/conftest.py
@@ -17,18 +17,6 @@ import logging
 
 import pytest
 
-# Metadata-bus layout constants — used by the response-word helpers below.
-# Mirrors the subset of constants imported by tests/protocols/test_rocev2.py
-# for the same helpers; co-locate the import so the helpers are self-contained
-# when multiple test files consume them via conftest.
-from pyrogue.protocols._RoCEv2 import (
-    _METADATA_PD_T,
-    _METADATA_MR_T,
-    _METADATA_QP_T,
-    _MR_KEY_B,
-    _IBV_QPS_RTS,
-)
-
 
 # ---------------------------------------------------------------------------
 # Re-export of wait_for + MemoryRoot from tests/conftest.py.
@@ -82,97 +70,3 @@ def rxe_server():
         except Exception as e:
             logging.getLogger("rocev2.teardown").warning(
                 "server.stop() failed during fixture teardown: %s", e)
-
-
-# ---------------------------------------------------------------------------
-# Metadata-bus response-word builders — shared by all tests in tests/protocols/.
-#
-# Each _pack_<type>_ok returns a 276-bit RX-bus word that satisfies the
-# corresponding _decode_<type>_resp in _RoCEv2.py. The _fail variants clear
-# ONLY the success bit, leaving every other field identical to the _ok word.
-# _pack_unknown_type emits a word whose resp_type tag is 3 (reserved) so the
-# "expected PD/MR/QP response, got type=3" branches in _roce_setup_connection
-# can be exercised without touching production code.
-#
-# These are NOT pytest fixtures — they are plain module-level functions.
-# Tests import them explicitly: `from conftest import _pack_pd_ok`.
-# (pytest adds the conftest.py directory to sys.path at collection time, so
-# the bare `conftest` import works without a package __init__.py.)
-# ---------------------------------------------------------------------------
-
-
-def _pack_pd_ok(pd_handler=0xDEADBEEF):
-    """RX word: PD response, success=1, pd_handler payload.
-
-    Matches _decode_pd_resp in _RoCEv2.py. resp_type tag at
-    bits [275:274]; success bit at bit 64; pd_handler at [63:32].
-    """
-    return ((_METADATA_PD_T & 0x3) << 274) | (1 << 64) | (pd_handler << 32)
-
-
-def _pack_mr_ok(lkey=0xABCD1234, rkey=0x5678DEAD):
-    """RX word: MR response, success=1, lkey/rkey payload.
-
-    Matches _decode_mr_resp in _RoCEv2.py. resp_type tag at
-    bits [275:274]; success bit at bit 262; lkey at [63:32]; rkey at [31:0].
-    """
-    return ((_METADATA_MR_T & 0x3) << 274) | (1 << 262) | (lkey << _MR_KEY_B) | rkey
-
-
-def _pack_qp_ok(qpn=0x123456, state=_IBV_QPS_RTS):
-    """RX word: QP response, success=1, qpn + qp_state payload.
-
-    Matches _decode_qp_resp in _RoCEv2.py. resp_type tag at
-    bits [275:274]; success bit at bit 273; qpn at [272:249];
-    qp_state at [216:213].
-    """
-    return ((_METADATA_QP_T & 0x3) << 274) | (1 << 273) | (qpn << 249) | (state << 213)
-
-
-# ---------------------------------------------------------------------------
-# Symmetric failure variants — identical to _ok except the success bit is
-# cleared.
-# ---------------------------------------------------------------------------
-
-
-def _pack_pd_fail(pd_handler=0xDEADBEEF):
-    """RX word: PD response, success=0, pd_handler payload preserved.
-
-    Identical to _pack_pd_ok except the success bit at bit 64 is 0.
-    Mirrors _roce_setup_connection's "FPGA PD allocation failed" branch.
-    """
-    return ((_METADATA_PD_T & 0x3) << 274) | (0 << 64) | (pd_handler << 32)
-
-
-def _pack_mr_fail(lkey=0xABCD1234, rkey=0x5678DEAD):
-    """RX word: MR response, success=0, lkey/rkey payload preserved.
-
-    Identical to _pack_mr_ok except the success bit at bit 262 is 0.
-    Simulates FPGA MR allocation failure (the 'FPGA MR allocation failed'
-    branch in _roce_setup_connection).
-    """
-    return ((_METADATA_MR_T & 0x3) << 274) | (0 << 262) | (lkey << _MR_KEY_B) | rkey
-
-
-def _pack_qp_fail(qpn=0x123456, state=_IBV_QPS_RTS):
-    """RX word: QP response, success=0, qpn + state payload preserved.
-
-    Identical to _pack_qp_ok except the success bit at bit 273 is 0.
-    """
-    return ((_METADATA_QP_T & 0x3) << 274) | (0 << 273) | (qpn << 249) | (state << 213)
-
-
-# ---------------------------------------------------------------------------
-# Reserved-tag builder — resp_type=3 is the unused 2-bit value (PD=0, MR=1,
-# QP=2 are the named values). Drives the "actual_type != _METADATA_*_T"
-# raise paths in _roce_setup_connection.
-# ---------------------------------------------------------------------------
-
-
-def _pack_unknown_type():
-    """RX word with resp_type tag = 3 (reserved / unused) and every other
-    field zeroed. Causes _decode_resp_type(rx) == 3 — every named decoder's
-    type-check branch in _roce_setup_connection raises rogue.GeneralError
-    with 'Expected PD/MR/QP response, got type=3'.
-    """
-    return (0x3 << 274)
